# contract/openapi.yaml
openapi: 3.1.0
info:
  title: AI 健身助手 P0 API
  version: 1.0.0
  description: >-
    微信首发但渠道无关的业务契约。关键训练数字由版本化确定性规则产生，
    AI 仅提供候选、排序和解释；用户归属始终由认证主体确定。
servers:
  - url: /
    description: local/test relative base
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/plans/{planId}/exercise-options:
    get:
      tags: [Plans]
      operationId: listPlanExerciseOptions
      description: 返回当前活动计划模板中可添加到指定训练日的规则处方；关键数字来自已验证模板，不由客户端或 AI 生成。
      parameters:
        - {name: planId, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: dayCode, in: query, required: true, schema: {type: string, minLength: 1, maxLength: 64, pattern: '^[A-Z0-9_]+$'}}
      responses:
        '200':
          description: 当前用户、器械、排除偏好和模板共同允许的动作处方
          content:
            application/json:
              schema: {$ref: './schemas/plan.yaml#/components/schemas/PlanExerciseOptionListResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/plans/{planId}/exercise-replacements:
    get:
      tags: [Plans]
      operationId: listPlanExerciseReplacements
      description: 返回指定计划日内原动作的已审核严格等价替换；候选保持原组数、次数和休息，仅在负重模式变化时要求重新校准。
      parameters:
        - {name: planId, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: dayCode, in: query, required: true, schema: {type: string, minLength: 1, maxLength: 64, pattern: '^[A-Z0-9_]+$'}}
        - {name: sourceExerciseCode, in: query, required: true, schema: {type: string, minLength: 1, maxLength: 64, pattern: '^[A-Z0-9_]+$'}}
      responses:
        '200':
          description: 同动作模式、主要肌群和难度且符合当前器械与排除偏好的替换候选；无合格项时 items 为空。
          content:
            application/json:
              schema: {$ref: './schemas/plan.yaml#/components/schemas/PlanExerciseReplacementOptionListResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/plans/{planId}/day-options:
    get:
      tags: [Plans]
      operationId: listPlanDayOptions
      description: 返回当前活动计划所属模板的完整训练日处方；客户端按当前工作副本过滤已使用项，关键数字来自已验证模板。
      parameters:
        - {name: planId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: 可用于恢复或添加到当前自定义计划中的模板训练日
          content:
            application/json:
              schema: {$ref: './schemas/plan.yaml#/components/schemas/PlanDayOptionListResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/DefaultError'}
...
  /api/v1/plans:
    post:
      tags: [Plans]
      operationId: createInitialPlan
      requestBody: {$ref: '#/components/requestBodies/PlanCreation'}
      responses:
        '201':
          description: 已激活候选计划；首次创建计划，已有计划时追加不可变的新版本
          content:
            application/json:
              schema: {$ref: './schemas/plan.yaml#/components/schemas/ActivePlanResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/plans/active:
    get:
      tags: [Plans]
      operationId: getActivePlan
      responses:
        '200':
          description: 当前活动计划及其不可变版本
          content:
            application/json:
              schema: {$ref: './schemas/plan.yaml#/components/schemas/ActivePlanResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/plans/{planId}/versions/{versionNo}:
    get:
      tags: [Plans]
      operationId: getPlanVersion
      parameters:
        - {$ref: '#/components/parameters/PlanIdPath'}
        - {$ref: '#/components/parameters/VersionNoPath'}
      responses:
        '200':
          description: 指定的历史计划版本
          content:
            application/json:
              schema: {$ref: './schemas/plan.yaml#/components/schemas/PlanVersionResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/plans/{planId}/versions:
    post:
      tags: [Plans]
      operationId: createPlanVersion
      parameters:
        - {$ref: '#/components/parameters/PlanIdPath'}
      requestBody: {$ref: '#/components/requestBodies/PlanVersionCreation'}
      responses:
        '200':
          description: 存在警告，必须携带 warningConfirmationToken 再次确认
          content:
            application/json:
              schema: {$ref: './schemas/plan.yaml#/components/schemas/PlanVersionResultResponse'}
        '201':
          description: 新不可变版本已创建并切换为活动版本
          content:
            application/json:
              schema: {$ref: './schemas/plan.yaml#/components/schemas/PlanVersionResultResponse'}
        '409': {$ref: '#/components/responses/VersionConflict'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/plans/{planId}/rebalance:
    post:
      tags: [Plans]
      operationId: previewPlanRebalance
      description: 只返回预览；USER_LOCKED 字段不得被静默覆盖。
      parameters:
        - {$ref: '#/components/parameters/PlanIdPath'}
      requestBody: {$ref: '#/components/requestBodies/PlanRebalance'}
      responses:
        '200':
          description: 重平衡预览，不创建版本
          content:
            application/json:
              schema: {$ref: './schemas/plan.yaml#/components/schemas/PlanVersionResultResponse'}
        '409': {$ref: '#/components/responses/VersionConflict'}
        default: {$ref: '#/components/responses/DefaultError'}
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
...
    PlanIdPath: {name: planId, in: path, required: true, schema: {type: string}}
    VersionNoPath: {name: versionNo, in: path, required: true, schema: {type: integer, minimum: 1}}
...
  requestBodies:
...
    PlanCreation: {required: true, content: {application/json: {schema: {$ref: './schemas/plan.yaml#/components/schemas/CreatePlanRequest'}}}}
    PlanVersionCreation: {required: true, content: {application/json: {schema: {$ref: './schemas/plan.yaml#/components/schemas/CreatePlanVersionRequest'}}}}
    PlanRebalance: {required: true, content: {application/json: {schema: {$ref: './schemas/plan.yaml#/components/schemas/RebalancePlanRequest'}}}}
...
  responses:
...
    BadRequest: {$ref: './schemas/common.yaml#/components/responses/BadRequest'}
    Unauthorized: {$ref: './schemas/common.yaml#/components/responses/Unauthorized'}
    NotFound: {$ref: './schemas/common.yaml#/components/responses/NotFound'}
    VersionConflict: {$ref: './schemas/common.yaml#/components/responses/VersionConflict'}
...
    DefaultError: {$ref: './schemas/common.yaml#/components/responses/DefaultError'}